Make a minimal effort to return correct-ish data
